Major overhaul of the feature to have the Manual Crop style used be setup to that the client only has to worry about the crop style that will be used by the site.
To do this setup a Adv Staff Display plugin that has both crop styles that can be used with a Feature Override that will switch the field and this plugin value to portrait from square.

Updated the views to have for each display version to have a square and portrait version.

Updated the Panels to have for each display option and square and portrait version.

Signed-off-by: Staris <jaison@worxco.com>